pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "cerebrium-hardware-store-agent"
version = "1.0.0"
description = "Hardware store voice AI agent built with LiveKit Agents, deployed on Cerebrium"
requires-python = ">=3.10, <3.14"

dependencies = [
    "livekit-agents[silero,turn-detector,openai]~=1.3",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "pyright",
]

[tool.pyright]
pythonVersion = "3.10"
include = ["src"]
typeCheckingMode = "basic"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "B", "A", "C4", "UP", "SIM", "RUF"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
